refact: Move router.js
